[NET-765] Replace deprecated methods and move some error classes
