+++ cmdb/framework/cmdb_object_manager.py
@@ -28,7 +28,7 @@ from bson import json_util
 from datetime import datetime
 from typing import List
 
-from cmdb.database.database_manager import PublicIDAlreadyExists
+from cmdb.database.errors.database_errors import PublicIDAlreadyExists
 from cmdb.event_management.event import Event
 from cmdb.framework.cmdb_base import CmdbManagerBase
 from cmdb.framework.managers.type_manager import TypeManager
